# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HotelBookingApi",
    "version": "1.0"
  },
  "paths": {
    "/api/Admin/seed": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Seed the database with test data.",
        "description": "This will populate the database with:\r\n- 10 Hotels with fake names and addresses\r\n- 6 Rooms per hotel (2 of each room type)\r\n- Sample bookings\r\n\r\nNote: \r\n* Room types (Single, Double, Deluxe) are created by database migrations.\r\n* If data already exists, seeding will be skipped.\r\n  Call the reset endpoint first if you want to start fresh.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Admin/reset": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Reset seeded test data",
        "description": "This will delete all test data from the database including bookings, rooms and hotels.\r\n\r\nNote: Room types are not deleted.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Bookings": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Create a new room booking",
        "requestBody": {
          "description": "Booking details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDataTransfer"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDataTransfer"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDataTransfer"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDataTransfer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Bookings/{reference}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get booking details by reference number",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "description": "Booking reference number",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDataTransfer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hotels": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Search for hotels by name",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "The hotel name to search for (partial match supported)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HotelDataTransfer"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hotels/{id}": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Get hotel details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The hotel ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelDetailDataTransfer"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hotels/{id}/rooms/available": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Find available rooms in a hotel for specified dates and guest count",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The hotel ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "checkIn",
            "in": "query",
            "description": "Check-in date (format: yyyy-MM-dd)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "checkOut",
            "in": "query",
            "description": "Check-out date (format: yyyy-MM-dd)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "guestCount",
            "in": "query",
            "description": "Number of guests",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "roomType",
            "in": "query",
            "description": "Optional room type filter (Single, Double, Deluxe)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailableRoomDataTransfer"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AvailableRoomDataTransfer": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "integer",
            "format": "int32"
          },
          "roomNumber": {
            "type": "string",
            "nullable": true
          },
          "roomTypeName": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "hotelId": {
            "type": "integer",
            "format": "int32"
          },
          "hotelName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookingResponseDataTransfer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "bookingReference": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "contactNumber": {
            "type": "string",
            "nullable": true
          },
          "guestCount": {
            "type": "integer",
            "format": "int32"
          },
          "checkInDate": {
            "type": "string",
            "format": "date"
          },
          "checkOutDate": {
            "type": "string",
            "format": "date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "room": {
            "$ref": "#/components/schemas/RoomDataTransfer"
          },
          "hotelName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateBookingDataTransfer": {
        "required": [
          "address",
          "checkInDate",
          "checkOutDate",
          "contactNumber",
          "firstName",
          "guestCount",
          "hotelId",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "hotelId": {
            "type": "integer",
            "format": "int32"
          },
          "roomType": {
            "type": "string",
            "description": "Optional preferred room type (Single, Double, Deluxe). If not specified, any available room will be allocated.",
            "nullable": true
          },
          "firstName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "lastName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "address": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "contactNumber": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "guestCount": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "checkInDate": {
            "type": "string",
            "format": "date"
          },
          "checkOutDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "A user-friendly error message.",
            "nullable": true
          },
          "error": {
            "type": "string",
            "description": "Technical error details (e.g., exception message).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Standard error response format for API errors."
      },
      "HotelDataTransfer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "totalRooms": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "HotelDetailDataTransfer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomDataTransfer"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoomDataTransfer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roomNumber": {
            "type": "string",
            "nullable": true
          },
          "roomTypeName": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "Admin"
    },
    {
      "name": "Bookings"
    },
    {
      "name": "Hotels"
    }
  ]
}